refactor `old` / `new` models
