+++ ocpmodels/models/__init__.py
@@ -6,13 +6,13 @@
 from .base import BaseModel  # noqa: F401
 from .cgcnn import CGCNN  # noqa: F401
 from .dimenet import DimeNetWrap as DimeNet  # noqa: F401
-from .dimenet_plus_plus import DimeNetPlusPlusWrap as DimeNetPlusPlus  # noqa: F401
+from .old_dimenet_plus_plus import DimeNetPlusPlusWrap as DimeNetPlusPlus  # noqa: F401
 from .fanet import FANet  # noqa: F401
-from .forcenet import ForceNet  # noqa: F401
+from .old_forcenet import ForceNet  # noqa: F401
 from .gemnet.gemnet import GemNetT  # noqa: F401
-from .new_dimenet_plus_plus import NewDimeNetPlusPlus  # noqa: F401
-from .new_forcenet import NewForceNet  # noqa: F401
-from .new_schnet import NewSchNet  # noqa: F401
-from .schnet import SchNetWrap as SchNet  # noqa: F401
+from .dimenet_plus_plus import NewDimeNetPlusPlus  # noqa: F401
+from .forcenet import NewForceNet  # noqa: F401
+from .schnet import NewSchNet  # noqa: F401
+from .old_schnet import SchNetWrap as SchNet  # noqa: F401
 from .sfarinet import SfariNet  # noqa: F401
 from .spinconv import spinconv  # noqa: F401
